Move abandoned cart into WooCommerce vendor space
[MAILPOET-4934]
