+++ mailpoet/assets/js/src/automation/integrations/mailpoet/index.tsx
@@ -2,7 +2,6 @@ import { registerStepType } from '../../editor/store';
 import { step as SendEmailStep } from './steps/send_email';
 import { step as SomeoneSubscribesTrigger } from './steps/someone-subscribes';
 import { step as WpUserRegisteredTrigger } from './steps/wp-user-registered';
-import { step as AbandonedCartTrigger } from './steps/abandoned-cart';
 import { step as AddTagsAction } from './steps/add_tags';
 import { step as RemoveTagsAction } from './steps/remove_tags';
 import { step as AddToListStep } from './steps/add_to_list';
@@ -12,7 +11,6 @@ import { step as UnsubscribeStep } from './steps/unsubscribe';
 import { step as NotificationEmail } from './steps/notification_email';
 import { registerStepControls } from './step-controls';
 import { registerAutomationSidebar } from './automation-sidebar';
-import { MailPoet } from '../../../mailpoet';
 
 export const initialize = (): void => {
   registerStepType(SendEmailStep);
@@ -27,9 +25,4 @@ export const initialize = (): void => {
   registerStepType(NotificationEmail);
   registerStepControls();
   registerAutomationSidebar();
-
-  if (!MailPoet.isWoocommerceActive) {
-    return;
-  }
-  registerStepType(AbandonedCartTrigger);
 };
