+++ mailpoet/assets/js/src/automation/integrations/woocommerce/index.tsx
@@ -1,5 +1,6 @@
 import { registerStepType } from '../../editor/store';
 import { step as OrderStatusChanged } from './steps/order_status_changed';
+import { step as AbandonedCartTrigger } from './steps/abandoned-cart';
 import { MailPoet } from '../../../mailpoet';
 
 export const initialize = (): void => {
@@ -7,4 +8,5 @@ export const initialize = (): void => {
     return;
   }
   registerStepType(OrderStatusChanged);
+  registerStepType(AbandonedCartTrigger);
 };
