+++ mailpoet/assets/js/src/automation/integrations/woocommerce/steps/abandoned-cart/icon.tsx
