+++ mailpoet/lib/Automation/Integrations/MailPoet/MailPoetIntegration.php
@@ -7,7 +7,6 @@ use MailPoet\Automation\Engine\Registry;
 use MailPoet\Automation\Integrations\MailPoet\Actions\SendEmailAction;
 use MailPoet\Automation\Integrations\MailPoet\Hooks\AutomationEditorLoadingHooks;
 use MailPoet\Automation\Integrations\MailPoet\Hooks\CreateAutomationRunHook;
-use MailPoet\Automation\Integrations\MailPoet\Subjects\AbandonedCartSubject;
 use MailPoet\Automation\Integrations\MailPoet\Subjects\SegmentSubject;
 use MailPoet\Automation\Integrations\MailPoet\Subjects\SubscriberSubject;
 use MailPoet\Automation\Integrations\MailPoet\SubjectTransformers\OrderSubjectToSegmentSubjectTransformer;
@@ -24,9 +23,6 @@ class MailPoetIntegration implements Integration {
 
   /** @var SubscriberSubject */
   private $subscriberSubject;
-
-  /** @var AbandonedCartSubject */
-  private $abandonedCartSubject;
 
   /** @var SomeoneSubscribesTrigger */
   private $someoneSubscribesTrigger;
@@ -53,7 +49,6 @@ class MailPoetIntegration implements Integration {
     ContextFactory $contextFactory,
     SegmentSubject $segmentSubject,
     SubscriberSubject $subscriberSubject,
-    AbandonedCartSubject $abandonedCartSubject,
     OrderSubjectToSubscriberSubjectTransformer $orderToSubscriberTransformer,
     OrderSubjectToSegmentSubjectTransformer $orderToSegmentTransformer,
     SomeoneSubscribesTrigger $someoneSubscribesTrigger,
@@ -65,7 +60,6 @@ class MailPoetIntegration implements Integration {
     $this->contextFactory = $contextFactory;
     $this->segmentSubject = $segmentSubject;
     $this->subscriberSubject = $subscriberSubject;
-    $this->abandonedCartSubject = $abandonedCartSubject;
     $this->orderToSubscriberTransformer = $orderToSubscriberTransformer;
     $this->orderToSegmentTransformer = $orderToSegmentTransformer;
     $this->someoneSubscribesTrigger = $someoneSubscribesTrigger;
@@ -82,7 +76,6 @@ class MailPoetIntegration implements Integration {
 
     $registry->addSubject($this->segmentSubject);
     $registry->addSubject($this->subscriberSubject);
-    $registry->addSubject($this->abandonedCartSubject);
     $registry->addTrigger($this->someoneSubscribesTrigger);
     $registry->addTrigger($this->userRegistrationTrigger);
     $registry->addAction($this->sendEmailAction);
