+++ mailpoet/lib/Automation/Integrations/WooCommerce/WooCommerceIntegration.php
@@ -3,6 +3,7 @@
 namespace MailPoet\Automation\Integrations\WooCommerce;
 
 use MailPoet\Automation\Engine\Registry;
+use MailPoet\Automation\Integrations\WooCommerce\Subjects\AbandonedCartSubject;
 use MailPoet\Automation\Integrations\WooCommerce\Subjects\CustomerSubject;
 use MailPoet\Automation\Integrations\WooCommerce\Subjects\OrderStatusChangeSubject;
 use MailPoet\Automation\Integrations\WooCommerce\Subjects\OrderSubject;
@@ -12,6 +13,9 @@ class WooCommerceIntegration {
 
   /** @var OrderStatusChangedTrigger */
   private $orderStatusChangedTrigger;
+
+  /** @var AbandonedCartSubject */
+  private $abandonedCartSubject;
 
   /** @var OrderStatusChangeSubject */
   private $orderStatusChangeSubject;
@@ -27,12 +31,14 @@ class WooCommerceIntegration {
 
   public function __construct(
     OrderStatusChangedTrigger $orderStatusChangedTrigger,
+    AbandonedCartSubject $abandonedCartSubject,
     OrderStatusChangeSubject $orderStatusChangeSubject,
     OrderSubject $orderSubject,
     CustomerSubject $customerSubject,
     ContextFactory $contextFactory
   ) {
     $this->orderStatusChangedTrigger = $orderStatusChangedTrigger;
+    $this->abandonedCartSubject = $abandonedCartSubject;
     $this->orderStatusChangeSubject = $orderStatusChangeSubject;
     $this->orderSubject = $orderSubject;
     $this->customerSubject = $customerSubject;
@@ -45,6 +51,7 @@ class WooCommerceIntegration {
       return $this->contextFactory->getContextData();
     });
 
+    $registry->addSubject($this->abandonedCartSubject);
     $registry->addSubject($this->orderSubject);
     $registry->addSubject($this->orderStatusChangeSubject);
     $registry->addSubject($this->customerSubject);
